Request persistent storage access, update source code
Otherwise data can get cleared if the browser decides there's not enough space, without the using being asked. This is pretty rare normally, but apparently Safari with cross-platform tracking protection on will delete data "created from scripts" after 7 days of not using the website... (image if it did that for every native app!). I'm not sure if this extends to "homescreens apps" as they call them, and if it's all data or what. But I think it's a good idea to persist anyway, as it's not stored anywhere else so would be lost otherwise.

Firefox shows a user prompt, Chromium browsers and Safari auto-accept or decline based on usage heuristics. So far Chrome is always declining on desktop despite having used the website a lot. Maybe because I'm prompting on page load, idk. I can't really do anything else as the JS and Lua bridging script doesn't work on Github pages so I'd have no idea when to prompt for permission (except a timeframe maybe..). But given it's a game I think it should be reasonably obvious why it wants persistent storage. Even if the browser auto-accept/decline doesn't realise that... 

Output of persistent storage permission is printed in console.