Remove purpose "any" from icons in manifest
This was causing Chromium desktop to always use the monochrome icon reason rather than regular one, I guess as it just saw "any" (even when placed after monochrome) and went with it. Now on Linux it seems to use the regular one for install UI, when hovering over the taskbar icon, in start menu, and loading mouse icon, but monochrome when not hovering over taskbar and in search, which is kinda bizarre but fine I think.