Output fullscreen support because Safari
It's still showing the fullscreen button and pressing the button is still trying to do something - despite the site now having updated and not throwing any errors about the feature checking code now. My theory is that Safari iOS claims to support requestFullscreen even though it doesn't actually.. would hardly be the first time Apple have done this (indexedDB, web Push..). Especially as they tried to implement fullscreen in iOS 17.6 but gave up, and are now trying to in iOS 18 beta (although off by default and doesn't seem to do anything when on).