Recompress screenshots at 90% quality
It's really hard to make out a difference when flicking back and forth on my laptop, except darth maul who's face is slightly blurry bizarrely. It reduces file size by 74% on average though, down 4.42 MiB from 5.52 to 1.40 of screenshots.

I would also have liked to shrink the resolution as on desktop they only ever take up a small portion o the screen, but on mobile they see to be showing at 1080p, so the difference would be slightly noticeable there.

Also worth noting that screenshots are loaded in browser which support rich install prompts (Chromium-based), but ignored in ones which don't (Firefox, Safari), which is good.