Preload game.data and store in cache rather than indexedDB
When the data is stored in indexedDB, the browser doesn't realise that it's already loaded, so would request it from the server on every (cached) load. Switching to storing game.data in cache - which I wanted to do anyway - means it is recognised, and also fixes https://github.com/FormularSumo/Star-Wars-Galaxy-Collection-Web/issues/4. In doing so I've also been able to almost half the size of game.js by removing all the caching and database code, making it much more readable.

I tried preloading love.wasm as well but doesn't really make any difference. The longest wait is always for game.data so loading that first is all that's needed. Ideally I'd like to defer loading screenshots till after the game I think but I've not figured out how to do that (if it's even possible). I guess one extreme approach would be to link to the manifest after game.data has been fetched but that seems a bit extreme.

Also fix some blankline/indent issues.